GraphDB_Pipeline/requirements.txt
aenum==3.1.12
aiohttp==3.8.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
charset-normalizer==2.1.1
frozenlist==1.3.3
gremlinpython==3.6.2
idna==3.4
isodate==0.6.1
multidict==6.0.4
nest-asyncio==1.5.6
psycopg2-binary==2.9.5
six==1.16.0
yarl==1.8.2
